Scripts/requirements.txt
mysql-connector-python==8.0.11
protobuf==3.5.2.post1
six==1.11.0

